Update Swift package for release
Replace `Package.swift` to build the package from `./Sources` (generated Swift bridge code)
and the binaries built from the previous commit and pushed as a "release" in the repo.

This completes the migration of the repo from its former use
as a simple package spec for the legacy `mobile_wallet` library
to the new one which is embedded in the repo.

Future releases are to follow the same two-step process layed out by this commit and its predecessor;
i.e. two conscutive, tagged commits within a single PR.
The complete process is described in the README.