Restructure: Move tests and rename library
A bunch of hideous tests are going to be added in an upcoming PR. Moving the test to their own file/module reduces the noise in the main file. The tests are entirely unmodified.

This separation is usually associated with integration tests in Rust. As the tests actually just exercise the exposed functions (which now actually have to be public), this seems appropriate.

It also seems appropriate in this regard to give the crate a less generic name than 'crypto'. It therefore is now named 'concordium-wallet-crypto-uniffi', reflecting that it should be able to target all languages supported by UniFFI, not just Swift (i.e. at least also Kotlin and Python). This has not been verified though.

Finally, the UniFFI dependency was bumped from v0.25.x to the latest version v0.26.x.