SwiftPM: Add ability to use local framework
Added processing of build-time environment variable `CONCORDIUM_WALLET_CRYPTO_FRAMEWORK_PATH` which may be used to select a locally compiled framework instead of the released one. This allows one to use an unreleased version during development of the SDK.

Note that even when used from downstream projects in other directories, the provided path is evaluated relative to the root of this project.