CI: Run tests on all platforms
The library is not Mac/iThing specific. UniFFI supports other languages that work on all platforms. Furthermore, Swift actually supports Linux and we might want the SDK to support that platform one day.

It therefore seems reasonable to avoid unknowingly adding anything that breaks Linux support in this library. Hence we should also test the Rust code on Windows and Ubuntu.