Name generated framework the same as package target
Rename generated framework to `ConcordiumWalletCryptoUniffi.xcframework` to match the (similarly renamed) binary target in `Package.swift`. We cannot just rename the target to match the existing generated file because that name (`ConcordiumWalletCrypto`) is already used by the library product target.

This is a requirement that apparently has been lifted in newer versions of Swift/macOS, but to support older versions, we follow it.