Add docstrings to Rust and UDL types and functions
In the Rust code, the docstrings explain the link between the types generated from UDL and the "original" ones from the core Concordium libraries that they bridge to. To properly refer to types located in concordium-base in the docstrings, an explicit dependency to that library has been added.

The docstrings on the UDL types are mostly what I could find in the docstrings on the original types.

As the dependency to concordium-base was added, the struct `Versioned` in `lib_test` was replaced by its original.