Update release instructions (#12)
Made release instructions clearer and change the "release" tag to happen after the release PR has been merged to `main`. This (along with the following) ensures that this PR isn't included in the release notes of the next *framework* release.

Added a workflow to publish a release on tags of the form `x.y.z` (i.e. for release of the Swift package). This means that we'll have releases both when we publish framework and package (the former now has "(framework)" appended to its release name). This workflow cannot be tested before it's been merged to `main`, but a `workflow_dispatch` trigger has been added in case we need to debug.

The old releases have already been updated to conform to the changes in this PR.

The indentation in the workflow files has been adjusted for consistency.

In the "Publish Binary Framework" workflow the `zip` command was replaced with `ditto` as that is recommended in various places because it's the command that Finder uses to compress. It apparently also includes things that `zip` doesn't which could be relevant if the contents were to be signed at some point...

Finally, the generated swift file was also renamed to `generated.swift` to prevent it from confusing reviewers to think that it's manual work 🙈