CI: Fix release workflow (#16)
The workflow (added in #12) erronously uses `github.ref` as the version name, but that value includes a `ref/tags/` prefix that should be removed. Added a step for doing that and ensured that the correct value is used.

Also needed to add submodule checkout for `cargo metadata` to work as well as throwing away logging output on stderr.

It was expected that there could be issues with this workflow becuse it couldn't be tested in advance. It's better tested now, but we still need a full successful run on `main` to be sure.