CI: Automatically trigger framework build by tag (#15)
To reduce the number of steps (and things to do wrong) in the release process, the workflow is updated to automatically trigger once a tag of form `build/*.*.*-*` is pushed. The workflow will then extract the version and run by itself.

A check that the version embedded in the tag actually matches the version of the crate was also included.